fixed an issue with lastProjectManager edit on contributorForm and factored out tables in invites
